feat: 페이지 생성 - profile - detected
