--- src/app/me/profile/layout.tsx
@@ -1,11 +1,0 @@
-export default function ProfileLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <>
-      <div className={"pb-[124px] pt-[72px]"}>{children}</div>
-    </>
-  );
-}
